+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FilmReviewDaoImpl.java
@@ -17,37 +17,37 @@ public class FilmReviewDaoImpl implements FilmReviewStorage {
     }
 
     @Override
-    public void addLikeToReview(Integer reviewId, Long userId) {
+    public void addLikeToReview(Long reviewId, Long userId) {
         jdbcTemplate.update("MERGE INTO film_reviews (review_id, user_id, is_like) VALUES (?, ?, ?)",
                 reviewId, userId, Boolean.TRUE);
         updateUsefulnessScore(reviewId);
     }
 
     @Override
-    public void deleteLikeFromReview(Integer reviewId, Long userId) {
+    public void deleteLikeFromReview(Long reviewId, Long userId) {
         jdbcTemplate.update("DELETE FROM film_reviews WHERE review_id=? AND user_id=?");
         updateUsefulnessScore(reviewId);
     }
 
     @Override
-    public void addDislikeToReview(Integer reviewId, Long userId) {
+    public void addDislikeToReview(Long reviewId, Long userId) {
         jdbcTemplate.update("MERGE INTO film_reviews (review_id, user_id, is_like) VALUES (?, ?, ?)",
                 reviewId, userId, Boolean.FALSE);
         updateUsefulnessScore(reviewId);
     }
 
     @Override
-    public void deleteDislikeFromReview(Integer reviewId, Long userId) {
+    public void deleteDislikeFromReview(Long reviewId, Long userId) {
         jdbcTemplate.update("DELETE FROM film_reviews WHERE review_id=? AND user_id=?");
         updateUsefulnessScore(reviewId);
     }
 
-    private void updateUsefulnessScore(Integer reviewId) {
+    private void updateUsefulnessScore(Long reviewId) {
         int usefulnessScore = getUsefulnessScore(reviewId);
         jdbcTemplate.update("UPDATE reviews SET useful=? WHERE review_id=?", usefulnessScore, reviewId);
     }
 
-    private int getUsefulnessScore(Integer reviewId) {
+    private int getUsefulnessScore(Long reviewId) {
         return jdbcTemplate.query("SELECT SUM(CASE WHEN is_like = TRUE THEN 1 ELSE -1 END) useful " +
                         "FROM film_reviews WHERE review_id =?", this::makeFilmReview, reviewId)
                 .stream().findAny().orElse(0);
